requirements.txt
appdirs==1.4.4
asgiref==3.3.1       
beautifulsoup4==4.9.3
bson==0.5.10
cachelib==0.1.1
certifi==2020.12.5
chardet==4.0.0
click==7.1.2
cycler==0.10.0
distlib==0.3.1
Django==3.1.7
django-crispy-forms==1.11.2
djangorestframework==3.12.4
filelock==3.0.12
Flask==1.1.2
Flask-Session==0.3.2
greenlet==1.1.0
gunicorn==20.1.0
idna==2.10
itsdangerous==1.1.0
Jinja2==2.11.3
joblib==1.0.1
kiwisolver==1.3.1
lxml==4.6.3
MarkupSafe==1.1.1
matplotlib==3.4.1
numpy==1.20.1
opencv-python==4.5.1.48
pandas==1.2.3
Pillow==8.1.2
psycopg2==2.8.6
pymongo==3.11.3
pyparsing==2.4.7
pytesseract==0.3.7
python-dateutil==2.8.1
pytz==2021.1
requests==2.25.1
scikit-learn==0.24.1
scipy==1.6.2
six==1.15.0
sklearn==0.0
soupsieve==2.2.1
SQLAlchemy==1.4.15
sqlparse==0.4.1
threadpoolctl==2.1.0
urllib3==1.26.4
virtualenv==20.4.2
Werkzeug==1.0.1